trilby-cli: extract Host to separate module
